Working on hammer evaluation in tactic_benchmark.lean
Currently, hammerBenchmarkFromModule can be used to limited success to
evaluate the hammer tactic given a recommendation from a JSON file, but
the code is not trustworthy enough to use for a full evaluation yet.
There are some known issues pertaining to how the JSON entry
corresponding to a particular constant is found, and there are also
discrepancies between how the hammer tactic actually performs and how
hammerBenchmarkFromModule says the hammer tactic performs.

To give one concrete example (though there are many), hammer succeeds on
ite_left in Mathlib.Data.Set.Basic.lean when given the recommendation
from the corresponding JSON file, but hammerBenchmarksFromModule
indicates that hammer fails on that theorem (and similar theorems).