Fix notation and namespace issues relating to premise selection
Note: Although progress has been made on the issue of introducing the
correct number of universal binders, this issue has not 100% been
resolved. Currently, the data extraction script
(training_data_with_premises), useHammerCore, and useHammer all agree on
the same approximation of how many of a theorem's universal binders
should be introduced. However, this approximation does not perfectly
replicate the number of universal binders that will have actually been
introduced when using "by hammer" on a theorem in the source code.